# openapi.json
{"openapi":"3.1.0","info":{"title":"API de Gestión de Envíos","description":"API REST para gestionar envíos - Taller de Integración de Sistemas","version":"1.0.0"},"paths":{"/":{"get":{"tags":["Root"],"summary":"Root","description":"Endpoint raíz que proporciona información básica de la API","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"tags":["Health"],"summary":"Health Check","description":"Endpoint de salud para verificar que el servicio está funcionando correctamente.\nÚtil para monitoreo y despliegues automatizados.","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/envios":{"get":{"tags":["Envíos"],"summary":"Obtener Envios","description":"Retorna la lista completa de envíos almacenados en el sistema","operationId":"obtener_envios_envios_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/Envio"},"type":"array","title":"Response Obtener Envios Envios Get"}}}}}},"post":{"tags":["Envíos"],"summary":"Crear Envio","description":"Crea un nuevo envío en el sistema\n\n- **id**: Identificador único del envío\n- **destinatario**: Nombre del destinatario\n- **direccion**: Dirección de entrega\n- **estado**: Estado actual del envío (Pendiente, En tránsito, Entregado)","operationId":"crear_envio_envios_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Envio"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Envio"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/envios/{id}":{"get":{"tags":["Envíos"],"summary":"Obtener Envio Por Id","description":"Retorna un envío específico basado en su ID\n\n- **id**: Identificador único del envío","operationId":"obtener_envio_por_id_envios__id__get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Envio"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Envio":{"properties":{"id":{"type":"integer","title":"Id"},"destinatario":{"type":"string","title":"Destinatario"},"direccion":{"type":"string","title":"Direccion"},"estado":{"type":"string","title":"Estado"}},"type":"object","required":["id","destinatario","direccion","estado"],"title":"Envio","example":{"destinatario":"Juan Pérez","direccion":"Av. República 123, Quito","estado":"En tránsito","id":1}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}